refactor: change filter file name
